+++ lib/components/home/home_description.dart
@@ -4,6 +4,7 @@ import "package:get/get.dart";
 import "package:x_containers/x_containers.dart";
 
 import "../../widgets/text/auto_color_text.dart";
+import "social_media_links.dart";
 
 /// A widget describing what the site is about.
 class HomeDescription extends StatelessWidget {
@@ -22,9 +23,15 @@ class HomeDescription extends StatelessWidget {
   Widget build(BuildContext context) {
     return XContainer(
       padding: EdgeInsets.all(XLayout.paddingL),
-      child: AutoColorText(
-        "home_description".tr,
-        style: context.textTheme.titleMedium,
+      child: Column(
+        children: [
+          AutoColorText(
+            "home_description".tr,
+            style: context.textTheme.titleMedium,
+          ),
+          XLayout.verticalM,
+          SocialMediaLinks(),
+        ],
       ),
     );
   }